Create Bank calculation.java
The Government of India recently demonetized the existing currency notes of 1000 and 500 denominations and issued new currency notes instead in 2000 and 500 denominations. People holding the old currency notes needed to exchange them for the new currency notes at bank branches. For this homework, we will assume that the bank will use the least number of notes principle, i.e., it will first calculate the 2000 denomination notes first, and the remaining amount, if any, as 500 denomination notes

Sample Console Output: