Create EulerConstant.java
Program the computation of the Euler constant using a loop for a given value of n
Write the static method, compute, that takes the number n as its argument. The method returns the computed result. Reuse the factorial computed in the previous iteration for the next iteration, i.e., do not explicitly compute each factorial in each iteration of the loop. In each iteration, print the intermediate results