Create BankingProceduralStyle_Parat1.java
Assume that the Certificates of Deposits (CDs) offered by various banks are compounded annually. The final value of the CD is computed using the following formula: FinalValue = InitialDeposit * (1 + r)t,  where t is the number of years the CD is invested, and r is the interest rate in decimal form. If the interest rate is 4%, then r = 0.04. 
Create a “procedural style” program that meets the requirements below. Create a class named BankingProceduralStyle and test the final value of two different CDs as specified below. BankA offers 3 year CDs with an interest rate of 4%. BankB offers 4 year CDs with an interest rate of 3%. You have $5000 to invest and considering the above two options. Use the variables for initial deposit, bank name, bank interest, and bank term (# of years) – one set of variables for the first CD and a different set of variables for the second CD.